+++ pkg/api/v1/function_resources.go
@@ -1,0 +1,54 @@
+///////////////////////////////////////////////////////////////////////
+// Copyright (c) 2017 VMware, Inc. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+///////////////////////////////////////////////////////////////////////
+
+package v1
+
+import (
+	strfmt "github.com/go-openapi/strfmt"
+
+	"github.com/go-openapi/errors"
+	"github.com/go-openapi/swag"
+)
+
+// NO TESTS
+
+// FunctionResources function resources
+// swagger:model FunctionResources
+type FunctionResources struct {
+
+	// cpu
+	CPU string `json:"cpu,omitempty"`
+
+	// memory
+	Memory string `json:"memory,omitempty"`
+}
+
+// Validate validates this function resources
+func (m *FunctionResources) Validate(formats strfmt.Registry) error {
+	var res []error
+
+	if len(res) > 0 {
+		return errors.CompositeValidationError(res...)
+	}
+	return nil
+}
+
+// MarshalBinary interface implementation
+func (m *FunctionResources) MarshalBinary() ([]byte, error) {
+	if m == nil {
+		return nil, nil
+	}
+	return swag.WriteJSON(m)
+}
+
+// UnmarshalBinary interface implementation
+func (m *FunctionResources) UnmarshalBinary(b []byte) error {
+	var res FunctionResources
+	if err := swag.ReadJSON(b, &res); err != nil {
+		return err
+	}
+	*m = res
+	return nil
+}